requirements.txt
python-telegram-bot==20.3
chromadb==0.4.22
openai>=1.0.0
google-cloud-vision
python-dotenv
numpy==1.26.4
psutil
requests
pydantic>=2.3.0
pydantic-settings>=2.1.0
